backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.2.0
gunicorn==21.2.0
